F17 (part 1): Auto-lock + admin adjust/void of submitted entries with lot & ledger reconciliation (#69) (#80)
* F17 (part 1): auto-lock + admin adjust/void of submitted entries with lot and ledger reconciliation (#69)

Domain:
- ManagerAdjust now accepts Submitted, Locked, and ManagerAdjusted entries
  (spec §8.1 MVP: one adjust path pre- and post-lock), requires a reason,
  and snapshots the values it replaces into AdjustedFromJson (the audit
  trail until the audit-log slice). Void(reason) preserves the entry as
  Voided. Lock(at) stamps LockedAtUtc. EggLot.AdjustProduction re-states a
  lot: the sold/allocated portion is untouchable, available absorbs the
  delta, 0 empties an unsold lot in place.
- EggLot.DailyEntryId links every lot to the entry whose submit generated
  it. (flock, date) is NOT an entry surface — multiple houses mean sibling
  entries per flock and day, and live verification caught an adjust zeroing
  a sibling entry lot. Migration backfills the link where exactly one
  non-draft entry matches; unprovable lots stay null and their entries
  refuse adjust/void (DailyEntry.PredatesLotTracking) rather than guess.

Application:
- AdjustDailyEntryHandler: client base Version (409 on mismatch, EF token
  backstop), flock lifecycle gate, grade validation (active+saleable for
  new ids; ids already on the entry grandfathered), then in ONE transaction:
  FOR UPDATE on the entry lots in the canonical (ProductionDate, Id)
  order shared with confirm/void, lots grown/shrunk/added/emptied to match
  the new lines — shrinking below a lot sold quantity is a 422 naming the
  grade — and a compensating bird movement for any mortality delta (more
  deaths -> Mortality row; fewer -> negative Adjustment; originals never
  edited).
- VoidDailyEntryHandler: same shape; empties every linked lot (422 if any
  eggs were sold), appends the reversing movement, preserves the entry.
- DailyEntryLockSweep: submitted entries older than 7 farm-local days
  (account timezone) lock automatically; runs per-account inside the
  DurableJobWorker guarded poll; idempotent and Version-race safe.

API: POST /daily-entries/{id}/adjust and /void, both AdminOnly (#73);
entry responses now carry version, adjustReason, voidReason, lockedAtUtc,
and the adjustedFrom snapshot.

Tests (222 green): lot reconciliation matrix (grow/shrink/add/remove),
shrink-below-sold 422 with full rollback, mortality compensation both
directions with CurrentBirds assertions, void paths, sibling-entry
isolation (the live-verify regression), parallel adjusts (exactly one
wins, Version delta 1), draft/stale/blank-reason guards, lock sweep, and
the worker-403 matrix extended with both new routes.

* Address codex review of PR #80

- Duplicate-grade lots: every lot keeps its sold floor and the first in
  canonical order carries the remainder — a later duplicate holding sold
  eggs no longer produces a spurious 422 when the aggregate target covers
  all sold eggs. Explicit aggregate-level 422 when it does not.
- Ledger notes truncate to BirdMovement.MaxNoteLength: a max-length reason
  no longer throws (500) when prefixed into the compensating movement note.
- Lock sweep uses a strict comparison — an entry exactly 7 farm-local days
  old keeps its final editable day; only strictly older entries lock.
- Loss-count sum validated as long in BOTH the adjust and record
  validators: three ints could overflow past a positive total and slip
  through the domain guard when no grade lines were present.
- AdjustedFromJson is plain text, not jsonb (provider-portability rule);
  the two migrations squashed into one clean AddDailyEntryAdjustments with
  the unambiguous lot backfill.
- Docs-sync in THIS PR: Help page (submit bullet, 7-day lock bullet, two
  Fixing-mistakes rows, three glossary rows) + GLOSSARY.md full lifecycle
  rewrite (Locked/ManagerAdjusted/Voided, lot linkage, sold-untouchable).
- New race test: parallel adjust vs void on the same base version —
  exactly one wins, Version delta 1, no blended lot state.

* Address pi review of PR #80

- Lock sweep isolates account failures: one account's bad timezone or
  transient DB error logs and moves on instead of starving the remaining
  accounts of their sweep. Shutdown cancellation still propagates.
- The success log reports the actually-locked count, not the batch size.

---------

+++ src/Cluckwork.Application/Features/DailyEntries/RecordDailyEntry/RecordDailyEntryValidator.cs
@@ -20,8 +20,10 @@ public sealed class RecordDailyEntryValidator : AbstractValidator<RecordDailyEnt
         RuleFor(x => x.DirtyEggs).GreaterThanOrEqualTo(0);
         RuleFor(x => x.DiscardedEggs).GreaterThanOrEqualTo(0);
         RuleFor(x => x.MortalityCount).GreaterThanOrEqualTo(0);
+        // long accumulation — three ints can overflow past a positive total
+        // (same fix as the adjust validator, codex review of PR #80).
         RuleFor(x => x)
-            .Must(x => x.CrackedEggs + x.DirtyEggs + x.DiscardedEggs <= x.TotalEggs)
+            .Must(x => (long)x.CrackedEggs + x.DirtyEggs + x.DiscardedEggs <= x.TotalEggs)
             .WithName("Eggs")
             .WithMessage("Cracked + dirty + discarded cannot exceed total eggs.");
